requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
pymongo>=4.6.0
requests>=2.31.0
python-dotenv>=1.0.0
flasgger>=0.9.7

# Package modèle (dossier frère ``../phosalert-model``). Depuis phosalert-backend :
#   pip install -e ../phosalert-model
-e ../phosalert-model

# Script Copernicus ``scripts/fetch_gabes_water.py`` (optionnel)
-r requirements/gabes-water.txt
